jpype second attempt
