+++ src/main/python/nutmeg_reader/nutmeg_reader.py
@@ -5,15 +5,50 @@ import magic
 import warnings
 
 from typing import Union
+from pkg_resources import get_distribution
 
 import h5py as h5
 import pandas as pd
 
-from jnius import autoclass
+import jpype
+import jpype.imports
+from jpype.types import *
 
-NutmegRealPlot      = autoclass('edlab.eda.reader.nutmeg.NutmegRealPlot')
-NutmegComplexPlot   = autoclass('edlab.eda.reader.nutmeg.NutmegComplexPlot')
-NutReader           = autoclass('edlab.eda.reader.nutmeg.NutReader')
+#from jnius import autoclass
+
+#NutmegRealPlot      = autoclass('edlab.eda.reader.nutmeg.NutmegRealPlot')
+#NutmegComplexPlot   = autoclass('edlab.eda.reader.nutmeg.NutmegComplexPlot')
+#NutReader           = autoclass('edlab.eda.reader.nutmeg.NutReader')
+
+__name__    = 'nutmeg_reader'
+__version__ = get_distribution(__name__).version
+
+
+def __get_maven_home():
+    mvn_command  = 'mvn help:evaluate -Dexpression=settings.localRepository -B 2> /dev/null'
+    maven_output = os.popen(mvn_command).read()
+    maven_home   = list(filter( lambda l: (not l.startswith('[')) and l
+                              , maven_output.split(sep='\n')))[0]
+    return maven_home
+
+def __default_class_path():
+    maven_home = __get_maven_home()
+    class_path = f'{maven_home}/edlab/eda/reader.nutmeg/{__version__}/reader.nutmeg-{__version__}-jar-with-dependencies.jar'
+    return class_path
+
+def __import_nutmeg():
+    global NutReader, NutmegRealPlot, NutmegComplexPlot
+    from edlab.eda.reader.nutmeg import NutReader, NutmegRealPlot, NutmegComplexPlot
+
+try:
+    jpype.startJVM()
+    __import_nutmeg()
+except ModuleNotFoundError as mnfe:
+    warnings.warn( message=f'Imporing Nutmeg with default class path failed.'
+                 , category=ImportWarning
+                 , )
+    jpype.addClassPath(__default_class_path())
+    __import_nutmeg()
 
 def _analysis_type(plot_name: str) -> str:
     analysis_pattern = "`(.*?)'"
